Rename entities.
